Finagling the concede button

+++ app/controllers/api/v1/games_controller.rb
@@ -2,7 +2,14 @@ class Api::V1::GamesController < ApplicationController
   def index
     games = current_user.games
     serialized_games = games.map do |game|
-      GameIndexSerializer.new(game)
+      match = game.matches.where(user: current_user)[0]
+      if match.endgame_confirm == false
+        GameIndexSerializer.new(game)
+      else
+      end
+    end
+    serialized_games = serialized_games.select do |match|
+      match != nil
     end
     render json: {
       currentUser: current_user,
@@ -14,7 +21,6 @@ class Api::V1::GamesController < ApplicationController
     data = JSON.parse(request.body.read)
     new_game = Game.new
     new_match = Match.new(user: User.find(data["id"]), game: new_game)
-    
     if new_game.save && new_match.save
       new_game = GameIndexSerializer.new(new_game)
       
@@ -25,19 +31,30 @@ class Api::V1::GamesController < ApplicationController
   end
   
   def show
+    
     current_game = Game.find(params["id"])
     gameState = ""
     opponent = nil
+    winner = nil
     cards = [].to_json
-    
     if (current_game.users.length == 2 && current_game.users.include?(current_user))
-      #THIS IS AN IN-PROGRESS GAME
-      gameState = "play"
-      user = current_user
-      opponent = current_game.users.where.not(username: current_user.username)
-      opponent = UserSerializer.new(opponent[0])
-      cards = Card.get_game_state(current_game)
-      whose_turn = UserSerializer.new(User.find(current_game.whose_turn_id))
+      #THIS IS AN IN-PROGRESS GAME OR A COMPLETED/CONCEDED GAME
+      if current_game.winner_id == nil
+        gameState = "play"
+        user = current_user
+        opponent = current_game.users.where.not(username: current_user.username)
+        opponent = UserSerializer.new(opponent[0])
+        cards = Card.get_game_state(current_game)
+        whose_turn = UserSerializer.new(User.find(current_game.whose_turn_id))
+      else
+        gameState = "complete"
+        user = current_user
+        opponent = current_game.users.where.not(username: current_user.username)
+        opponent = UserSerializer.new(opponent[0])
+        cards = nil
+        whose_turn = nil
+        winner = UserSerializer.new(User.find(current_game.winner_id))
+      end
     elsif (current_game.users.length == 1 && current_game.users[0] == current_user)
       #THIS IS A GAME IN WHICH USER IS WAITING FOR AN OPPONENT
       #THEY SHOULD ONLY HAVE THE OPTION TO DELETE OR GO BACK
@@ -64,21 +81,50 @@ class Api::V1::GamesController < ApplicationController
       #THIS GAME NO LONGER EXISTS ERROR
       gameState = "error"
     end
-      
+    
     render json: {
       gameState: gameState,
       currentUser: user,
       opponent: opponent,
       cards: cards,
       whose_turn: whose_turn,
-      card_reference: Card.all
+      card_reference: Card.all,
+      winner: winner
     }
   end
   
   def destroy
-    game_to_delete = Game.find(params["id"])
-    game_to_delete.destroy
-    render json: { response: "ok" }
+    winner = nil
+    if params["gameState"] == "concession"
+      game_to_concede = Game.find(params["id"])
+      user = params["user"]
+      winner = game_to_concede.users.where.not(id: user["id"])
+      game_to_concede.winner_id = winner[0].id
+      game_to_concede.whose_turn_id = nil
+      game_to_concede.gamestate = nil
+      game_to_concede.save
+      
+      
+      winner = UserSerializer.new(winner[0])
+    elsif params["gameState"] == "deleteWithoutLoss"
+      game_to_delete = Game.find(params["id"])
+      game_to_delete.destroy
+    elsif params["gameState"] == "confirmGameOver"
+      game_to_delete = Game.find(params["id"])
+      current_player_match = game_to_delete.matches.where(user_id: params["user"]["id"])[0]
+      current_player_match.endgame_confirm = true
+      current_player_match.save
+      
+      both_matches = game_to_delete.matches
+      if both_matches[0].endgame_confirm == true && both_matches[1].endgame_confirm == true
+        game_to_delete.destroy
+      end
+    end
+    
+    render json: { 
+      gameState: "done",
+      winner: winner
+    }
   end
   
   def update
